refactor frontend - change foldername and remove timer for showing skeleton in productslist
